implemented authentication and authorization

+++ src/OA.Infrastructure/Controllers/UserController.cs
@@ -1,10 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authentication.Cookies;
+using Microsoft.AspNetCore.Mvc;
 using OA.Contracts;
 using OA.Service.Abstraction;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using System.Security.Claims;
 
 namespace UserCRUD_demo_Project_.Controllers
 {
@@ -18,12 +18,16 @@ namespace UserCRUD_demo_Project_.Controllers
             _serviceManager = serviceManager;
         }
 
+
         //get all users
         [HttpGet]
         [Route("getall")]
         public IActionResult GetAll()
         {
-            return Ok(_serviceManager.UserService.GetAllUsers());
+            if (User.Identity.IsAuthenticated)
+                return Ok(_serviceManager.UserService.GetAllUsers());
+            else
+                return Unauthorized("you are not Authorized!:Please Login");
         }
 
         //get one user
@@ -31,8 +35,13 @@ namespace UserCRUD_demo_Project_.Controllers
         [Route("get")]
         public IActionResult GetUser(int Id)
         {
-            if (_serviceManager.UserService.GetUserByID(Id) != null)
-                return Ok(_serviceManager.UserService.GetUserByID(Id));
+            if (!User.Identity.IsAuthenticated)
+                return Unauthorized("You are not Authorized!:Please Login");
+            var user = _serviceManager.UserService.GetUserByID(Id);
+            if (user != null)
+            {   
+                    return Ok(_serviceManager.UserService.GetUserByID(Id));
+            }
             else
                 return NotFound("User doesn't exists!");
         }
@@ -41,20 +50,24 @@ namespace UserCRUD_demo_Project_.Controllers
         [HttpPost("add")]
         public IActionResult AddUser(UserDto user)
         {
-
+            if (!User.Identity.IsAuthenticated)
+                return Unauthorized("You are not Authorized to add a user!");
             return Ok(_serviceManager.UserService.AddUser(user));
         }
         //remove user
         [HttpDelete("{Id}")]
         public IActionResult RemoveUser(int Id)
         {
+            if (!User.Identity.IsAuthenticated)
+                return Unauthorized("You are not Authorized to remove a user!");
             return Ok(_serviceManager.UserService.RemoveUser(Id));
         }
         //update user
         [HttpPut("update")]
         public IActionResult UpdateUser(UserDto user)
         {
-
+            if (!User.Identity.IsAuthenticated)
+                return Unauthorized("You are not Authorized to update a user!");
             return Ok( _serviceManager.UserService.UpdateUser(user));
         }
